Support pbc depth-1
