+++ mlir/include/PBC/Transforms/Passes.td
@@ -67,6 +67,14 @@ def AvoidYMeasureOption : Option<
     /*default=*/"false",
     "Avoid Pauli-Y measurements for Clifford rotations. Rather than performing a Pauli-Y measurement for Clifford rotations (sometimes more costly), a Y state is used instead (requires Y state preparation).">;
 
+def OnlyDisjointQubitOption : Option<
+    /*C++ var name=*/"onlyDisjointQubit",
+    /*CLI arg name=*/"disjoint-qubit",
+    /*type=*/"bool",
+    /*default=*/"false",
+    "Group ops into the same layer only when they act on disjoint qubits. "
+    "By default, commuting ops on overlapping qubits may also be merged into one layer.">;
+
 //===----------------------------------------------------------------------===//
 //                               Passes
 //===----------------------------------------------------------------------===//
@@ -131,6 +139,8 @@ def CountPPMSpecsPass : Pass<"ppm-specs"> {
 
 def PartitionLayersPass : Pass<"partition-layers"> {
     let summary = "Partition layers of Pauli Product Measurement operations.";
+
+    let options = [OnlyDisjointQubitOption];
 }
 
 def TLayerReductionPass : Pass<"reduce-t-depth"> {
