+++ mlir/lib/PBC/Utils/PBCLayer.cpp
@@ -16,14 +16,59 @@
 
 #include "llvm/ADT/STLExtras.h"
 
+#include "PBC/IR/PBCOpInterfaces.h"
 #include "PBC/IR/PBCOps.h"
 #include "PBC/Utils/PauliStringWrapper.h"
 #include "Quantum/IR/QuantumOps.h" // for quantum.extract op
+
+#include <mlir/Support/WalkResult.h>
 
 using namespace catalyst::pbc;
 
 namespace catalyst {
 namespace pbc {
+
+bool isParentLayerOp(PBCOpInterface op)
+{
+    // Skip ops nested inside an existing pbc.layer region
+    auto parentOp = op->getParentOp();
+    while (parentOp != nullptr) {
+        if (isa<LayerOp>(parentOp))
+            return true;
+
+        parentOp = parentOp->getParentOp();
+    }
+
+    return false;
+}
+
+// Partition PBC ops into layer groups using commutation/disjoint-qubit rules.
+// Only op membership is recorded; operand/result bookkeeping is deferred until
+// construction so that SSA values reflect any layers already materialized in IR.
+llvm::SmallVector<std::vector<PBCOpInterface>>
+PBCLayerContext::groupLayers(mlir::Operation *root, bool onlyOnDisjointQubit)
+{
+    llvm::SmallVector<std::vector<PBCOpInterface>> groups;
+    PBCLayer layer(this);
+    root->walk([&](PBCOpInterface op) {
+        if (isParentLayerOp(op)) {
+            return WalkResult::skip();
+        }
+
+        if (layer.insert(op, onlyOnDisjointQubit)) {
+            return WalkResult::skip();
+        }
+
+        groups.emplace_back(layer.getOps());
+        layer = PBCLayer(this);
+        layer.insert(op);
+
+        return WalkResult::advance();
+    });
+
+    groups.emplace_back(layer.getOps());
+    return groups;
+}
 
 void PBCLayer::insertToLayer(PBCOpInterface op)
 {
@@ -182,7 +227,7 @@ bool PBCLayer::insertsAreAfterExistingOps(PBCOpInterface op) const
     return true;
 }
 
-bool PBCLayer::insert(PBCOpInterface op)
+bool PBCLayer::insert(PBCOpInterface op, bool onlyDisjointQubit)
 {
     if (empty()) {
         insertToLayer(op);
@@ -194,19 +239,19 @@ bool PBCLayer::insert(PBCOpInterface op)
     if (!isSameBlock(op) || !extractsAreBeforeExistingOps(op) || !insertsAreAfterExistingOps(op)) {
         return false;
     }
-
+    
     // 3. It acts on disjoint qubits
     // 4. Or it commutes with all the ops in the layer
     if (actOnDisjointQubits(op)) {
         insertToLayer(op);
         return true;
     }
-
-    if (commuteToLayer(op)) {
+    
+    // If onlyOnDisjointQubit is true, we only check the disjoint qubit condition
+    if (commuteToLayer(op) && !onlyDisjointQubit) {
         insertToLayer(op);
         return true;
     }
-
     return false;
 }
